+++ functions.php
@@ -540,14 +540,15 @@ final class OCEANWP_Theme_Class {
 		$vh_target   = get_theme_mod( 'ocean_vertical_header_dropdown_target', 'link' );
 		$vh_target   = $vh_target ? $vh_target : 'link';
 		$array       = array(
-			'isRTL'                => is_rtl(),
-			'menuSearchStyle'      => oceanwp_menu_search_style(),
-			'sidrSource'           => oceanwp_sidr_menu_source(),
-			'sidrDisplace'         => get_theme_mod( 'ocean_mobile_menu_sidr_displace', true ) ? true : false,
-			'sidrSide'             => $sidr_side,
-			'sidrDropdownTarget'   => $sidr_target,
-			'verticalHeaderTarget' => $vh_target,
-			'customSelects'        => '.woocommerce-ordering .orderby, #dropdown_product_cat, .widget_categories select, .widget_archive select, .single-product .variations_form .variations select',
+			'isRTL'                 => is_rtl(),
+			'menuSearchStyle'       => oceanwp_menu_search_style(),
+			'mobileMenuSearchStyle' => oceanwp_mobile_menu_search_style(),
+			'sidrSource'            => oceanwp_sidr_menu_source(),
+			'sidrDisplace'          => get_theme_mod( 'ocean_mobile_menu_sidr_displace', true ) ? true : false,
+			'sidrSide'              => $sidr_side,
+			'sidrDropdownTarget'    => $sidr_target,
+			'verticalHeaderTarget'  => $vh_target,
+			'customSelects'         => '.woocommerce-ordering .orderby, #dropdown_product_cat, .widget_categories select, .widget_archive select, .single-product .variations_form .variations select',
 		);
 
 		// WooCart.