+++ inc/helpers.php
@@ -1753,6 +1753,31 @@ if ( ! function_exists( 'oceanwp_menu_search_style' ) ) {
 }
 
 /**
+ * Returns mobile menu search style
+ *
+ * @since 2.0.9
+ */
+if ( ! function_exists( 'oceanwp_mobile_menu_search_style' ) ) {
+
+	function oceanwp_mobile_menu_search_style() {
+
+		// Get search style from Customizer
+		$style = get_theme_mod( 'ocean_mobile_menu_search_style', 'drop_down' );
+
+		// Apply filters for advanced edits
+		$style = apply_filters( 'oceanwp_mobile_menu_search_style', $style );
+
+		// Sanitize output so it's not empty and return
+		$style = $style ? $style : 'drop_down';
+
+		// Return style
+		return $style;
+
+	}
+
+}
+
+/**
  * Adds the search icon to the menu items
  *
  * @since 1.0.0
@@ -4698,3 +4723,111 @@ function owp_parse_attr( $context, $attributes = array(), $args = array() ) {
 	// Apply filter based on context.
 	return apply_filters( "owp_attr_{$context}", $attributes, $context, $args );
 }
+
+/**
+ * Search Icon
+ */
+function oceanwp_mobile_search_icon() {
+
+	$class        = '';
+	$search_style = oceanwp_mobile_menu_search_style();
+	$header_style = oceanwp_header_style();
+
+	if ( 'disabled' === $search_style || 'vertical' === $header_style ) {
+		return;
+	}
+
+	// Get correct search icon class
+	if ( 'drop_down' == $search_style ) {
+		$class = 'dropdown';
+	} elseif ( 'overlay' == $search_style ) {
+		$class = 'overlay';
+	} else {
+		$class = '';
+	}
+
+	?>
+
+	<a href="#" class="search-icon-<?php echo esc_attr( $class ); ?>"><?php oceanwp_icon( 'search' ); ?></a>
+
+	<?php
+}
+
+/**
+ * Search form
+ */
+function oceanwp_mobile_search_form_htlm() {
+
+	$class        = '';
+	$search_style = oceanwp_mobile_menu_search_style();
+	$header_style = oceanwp_header_style();
+	$post_type    = get_theme_mod( 'ocean_menu_search_source', 'any' );
+
+	if ( 'disabled' === $search_style || 'vertical' === $header_style ) {
+		return;
+	}
+
+	// Get correct search icon class
+	if ( 'drop_down' == $search_style ) {
+		$class = 'dropdown';
+	} elseif ( 'header_replace' == $search_style ) {
+		$class = 'overlay';
+	} elseif ( 'overlay' == $search_style ) {
+		$class = 'overlay';
+	} else {
+		$class = '';
+	}
+
+	?>
+
+	<div id="icon-searchform-<?php echo esc_attr( $class ); ?>" class="search-style-<?php echo esc_attr( $class ); ?>">
+		<?php
+		if ( 'drop_down' === $search_style ) { ?>
+			<form method="get" class="oew-searchform" action="<?php echo esc_url( home_url( '/' ) ); ?>">
+				<input type="text" class="field" name="s" placeholder="<?php oceanwp_theme_strings( 'owp-string-search-text', 'oceanwp' ); ?>">
+				<?php
+				if ( 'any' !== $post_type ) { ?>
+					<input type="hidden" name="post_type" value="<?php echo esc_attr( $post_type ); ?>">
+				<?php
+				} ?>
+			</form>
+		<?php
+		}
+
+		else if ( 'overlay' === $search_style ) { ?>
+			<div class="container clr">
+				<form method="get" class="oew-searchform" action="<?php echo esc_url( home_url( '/' ) ); ?>">
+					<a href="#" class="search-overlay-close"><span></span></a>
+					<input class="oew-search-overlay-input" type="search" name="s" autocomplete="off" value="">
+					<?php
+					if ( 'any' !== $post_type ) { ?>
+						<input type="hidden" name="post_type" value="<?php echo esc_attr( $post_type ); ?>">
+					<?php
+					} ?>
+					<label><?php oceanwp_theme_strings( 'owp-string-search-overlay-search-text', 'oceanwp' ); ?><span><i></i><i></i><i></i></span></label>
+				</form>
+			</div>
+		<?php
+		} ?>
+	</div>
+
+	<?php
+}
+
+/**
+ * Display search icon
+ */
+function mobile_menu_search_icon() {
+
+	$search_style = oceanwp_mobile_menu_search_style();
+	$header_style = oceanwp_header_style();
+
+	if ( 'disabled' === $search_style || 'vertical' === $header_style ) {
+		return;
+	}
+
+	add_action( 'ocean_after_mobile_icon', 'oceanwp_mobile_search_icon' );
+	add_action( 'ocean_mobile_menu_icon_after', 'oceanwp_mobile_search_form_htlm' );
+
+}
+add_action( 'wp', 'mobile_menu_search_icon' );